Don't validate element type in createElement/JSX
We validate these in the renderers inherently anyway and the renderers
are the only ones that really knows what a valid type for that renderer is.

For example the Client Reference check isn't valid for all kinds of configs
in Flight. Also, a lot of the checks for valid elements are not valid in Flight
e.g. class components.

The only reason to do this validation early is because it provides a better
stack trace of the callsite for context. However, that's a general problem
and the whole point of enableOwnerStacks that it solves this for all kinds of
late errors.

For now I keep the early validation when enableOwnerStacks is off because
in that case we don't have a good stack (but to be fair we don't for any
other error neither). Once it is further rolled out we can delete these fully.